Trying renaming of locusst file
